Add round-trip dump tool
This is a simple tool that ultimately does what the test code does (will
be added next commit), but doesn't do comparison or anything. It's
intended to be used with something like diff/colordiff so that you can
do large comparisons of the data coming out with the original golden
data.