state: Update StateModule references to be consistent across levels
The root module is a single-element reference so that it it can be
omitted if need be. This makes referencing it in callers awkward, as the
root will need to be a pointer value, but any sub-elements of the list
will need to be non-pointer values. This does not affect parsing in any
way, but makes consumption more predictable.