Custom unmarshaling for ConfigResource expressions
Expressions in the Terraform config JSON output can take one of two
forms:

* Singular, scalar expression types containing either a
"constant_value" or "references" field, or:
* A list of objects, each one representing a block of expressions.
This form is what shows up when the referenced field is a nested
block (sometimes referred to as a "sub-resource", although this is a
misnomer).

This cannot be represented in JSON using just a struct alone due to
the ambiguity as this ultimately means that you can have either an
Expression, or a []map[string]Expression. As such, custom parsing is
necessary to determine the exact shape of the data.

This adds that custom behavior to ConfigResource, by wrapping the
actual data in an embedded ConfigResourceData struct. We then
populate the expression list with a custom Unmarshaler. A custom
Marshaler reverses the process.

This ultimately puts the nested blocks in the new NestedBlocks field
in a particular Expression. This will still mean that only one of the
(now) three fields will be populate at any given time.

Note that the custom parsing behavior is only attached to
ConfigResource at this point in time as this is the only place that
nested expression blocks should show up, AFAIK. If this changes, this
this behavior will be extended to the places it needs to be as well.