Make all complex type references pointers
Unfortunately, the nature of the golang JSON library usually means
that if you want to omit a specific complex singular field, it needs
to be a pointer so that a "nil" denotes "empty" and omitted, versus a
non-nil reference to an empty denote not omitted, but minimal. Maps
and lists are exceptions to this where len == 0 can either mean
nil and non-nil but empty.

This was leading to many situations where singular fields for a type
were being referenced as *Type, and lists and maps were being
referenced as []Type, for example, causing inconsistencies and
leading to exceptions having to be made in the code (ie: to properly
implement Marshal for Expression, the MarshalJSON needed to be
a value versus pointer receiver).

Editing all the references so that they were consistently pointers
across the board has been something I've been thinking of doing for a
few days, so I've just done it.