config: Remove ProviderConfigKey parsing, document as opaque
The two references to provider config keys (one in the
ProviderConfigs field in the top-level Config struct, and the other
in an individual ConfigResource instance) should not be parsed and
are being marked as opaque in the specification. As such, they should
not be parsed and should be explicitly documented as not being
passable.

Further to that, the value in ConfigResource should not be relied to
to be able to find a resource in ProviderConfigs, as there are at
least two scenarios right now where this will provide undefined
behavior:

* The provider declaration is implicit, or:
* The provider is implicit and the ConfigResource in question belongs
to a resource within a module, in which case the module address is
always encoded into the key. This prevents even explicit, inherited
providers from being looked up on part that the config key is opaque.