[WIP] JSON Schema additions
This is starting the work for the JSON schema representation.

Most things have been added now, we just need to dive into
attributes, and then we should be set.

Note that I've asked about the possibility of filtering the returned
schema data to ensure that we only get the data we need, as the
output of "terraform providers schema -json" is currently quite
large.