Add support for object types
This adds support for object types.

Object types are new HCL2 types that are shimmed from the current SDK
from schema fields that are generally list or set type, with a max
length of 1 and a resource element type.

Output of object types in the schema JSON is new as of beta2, hence
why it was not added earlier.