Replace SchemaAttributeType et al with cty
This drop-in replaces the SchemaAttributeType types with cty.

As SchemaAttributeType et al were basically modeled after a subset of
cty, this was as simple as just setting AttributeType in
SchemaAttribute to be cty.Type instead of SchemaAttributeType. No
marshal/unmarshal tests needed to be altered and things pass as
expected.