Add support for format versions 1.0 and above (#41)
A future release of Terraform will ship with a 1.0 format version for
all stable JSON formats, which will be backwards compatible with the
existing format. This will also mark a release where the JSON
compatibility guarantee is more clearly defined, and so format versions
less than 2.0 should always be backwards compatible with consumers which
support 1.0.

This commit prepares for that change by adding support for version
constraints rather than supported fixed sets of versions. This is done
using the go-version package.

Includes test data from a pre-release of Terraform 1.1 which includes
the planned JSON format version changes.